PPU Working in Progress
Currently Pattern Tables are getting loaded

+++ c6502.cpp
@@ -1030,7 +1030,7 @@ std::map<uint16_t, std::string> c6502::disassemble(uint16_t nStart, uint16_t nSt
 		else if (lookup[opcode].addrmode == &c6502::REL)
 		{
 			value = bus->cpuRead(addr, true); addr++;
-			sInst += "$" + hex(value, 2) + " [$" + hex(addr + value, 4) + "] {REL}";
+			sInst += "$" + hex(value, 2) + " [$" + hex(addr + (int8_t)value, 4) + "] {REL}";
 		}
 
 		// Add the formed string to a std::map, using the instruction's
